test: split S8 provider scenario helpers
